experiments/DualViewHair/requirements.txt
# DualViewHair Dependencies

# Core ML frameworks
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0

# Computer vision
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Face detection and landmarks (optional)
dlib>=19.22.0  # Face landmarks
mediapipe>=0.8.0  # Alternative face detection

# Data processing
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0

# Configuration and logging
PyYAML>=5.4.0
tensorboard>=2.7.0
tqdm>=4.62.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Jupyter notebooks (for analysis)
jupyter>=1.0.0
ipywidgets>=7.6.0

# Development tools
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0

# Optional: Advanced features
# faiss-cpu>=1.7.0  # For fast similarity search
# umap-learn>=0.5.0  # For embedding visualization
# albumentations>=1.1.0  # Advanced augmentations
